Remove partial support for Index::$concurrent (#18802)
Remove partial support for Index::$concurrent

I had this half implemented because it was present in migrations. After
finishing the feature off, I decided against it. While having
`concurrent` would help migrations it cannot be reflected, and thus
would always be null except in the rare circumstances it is needed.
Instead we can have `migrations` do a `str_replace()` as required.